Played around with a default SVM classifier from sklearn.
I messed around in a sandbox notebook to get familiar with
the sklearn modules for SVM classification and cross validation.
I also used pandas and numpy to do some manipulation of the
original data to get it into separate csv files suitable for
the sklearn classifiers to import as numpy ndarrays.