Added fitness function code/documentation. Restructured data division.
This work was accidentally done in the master branch; it should have
been done in the pso_tests branch, but it doesn't appear to have
hurt anything. The two classes are nearly at a complete first draft.
There was some data restructuring. The inertia coefficient, w,
was moved from being stored as a global parameter in the COMB_Swarm
class in file cpso_swarm.py, to being stored individually
in each COMB_Particle class, in file cpso_particle.py. This is
because I received a new equation for w that is individual to
each particle. Additionally, I finished docstrings for everything
that has been written so far. The only functionality missing is
the ability for the fitness function to subset the feature
space.