Added enormous reporting functionality and a default wrapper script.
cpso_particle.py and cpso_swarm.py both got some new attributes
to facilitate in-progress reporting. cpso_swarm.py handles
most of the recording. Actual data export is handled by the
wrapper script. test_cpso_wrapper.py does an enormous amount
of work wrapping the two classes for the algorithm. It parses
all parameters as command line arguments, runs a single instance
of the algorithm with those parameters, and then exports a bunch
of data as well as a summary report log, all in the target output
directory. It does some basic error checking for the values given
in the command line parameters.