Took the wrapper script test_cpso_wrapper.py to full function.
Finished a first, functional draft of a wrapper script for the
COMB_Particle and COMB_Swarm classes that runs one single experiment
using files for data input, an output directory, and all command
line arguments and no hard coded parameters.

Note that the classifier and classifier scoring is all still
hard coded, but the algorithm parameters are all given on
the command line.