I forgot to commit for a few days, so this is a large complex one.
I finished debugging the Using COMB-PSO - First Run experiment and
ran it on Pegasus. I also had a meeting that resulted in me
abandoning the svm I had been using. So I mostly abandoned the
results from that experiment and immediately ran it again with
a KNN classifier instead. I also did a quick experiment comparing
the performance of different classifiers on this dataset. I finished
the documentation of Using COMB-PSO - First Run, but have not analyzed
or documented Using COMB-PSO - KNN yet.